add builder to template tests
