+++ docs/template_plugin/tests/functional/op_reference/acosh.cpp
@@ -14,42 +14,17 @@
 
 using namespace ngraph;
 
+namespace reference_tests {
 namespace {
-
-struct Tensor {
-    Tensor() = default;
-
-    Tensor(const ngraph::Shape& shape, ngraph::element::Type type, const InferenceEngine::Blob::Ptr& data): shape {shape}, type {type}, data {data} {}
-
-    template <typename T>
-    Tensor(const ngraph::Shape& shape, ngraph::element::Type type, const std::vector<T>& data_elements)
-        : Tensor {shape, type, CreateBlob(type, data_elements)} {}
-
-    ngraph::Shape shape;
-    ngraph::element::Type type;
-    InferenceEngine::Blob::Ptr data;
-};
 
 struct AcoshParams {
     Tensor input;
     Tensor expected;
 };
 
-struct Builder {
-    AcoshParams params;
-
-    operator AcoshParams() const {
-        return params;
-    }
-
-#define ADD_SET_PARAM(set_p)                   \
-    Builder& set_p(decltype(params.set_p) t) { \
-        params.set_p = std::move(t);           \
-        return *this;                          \
-    }
-    ADD_SET_PARAM(input);
-    ADD_SET_PARAM(expected);
-#undef ADD_SET_PARAM
+struct Builder : ParamsBuilder<AcoshParams> {
+    REFERENCE_TESTS_ADD_SET_PARAM(Builder, input);
+    REFERENCE_TESTS_ADD_SET_PARAM(Builder, expected);
 };
 
 class ReferenceAcoshLayerTest : public testing::TestWithParam<AcoshParams>, public CommonReferenceTest {
@@ -103,3 +78,4 @@ INSTANTIATE_TEST_SUITE_P(
                           .input({{8}, element::u64, std::vector<uint64_t> {1, 2, 3, 4, 5, 10, 100, 1000}})
                           .expected({{8}, element::u64, std::vector<uint64_t> {0, 1, 2, 2, 2, 3, 5, 8}})),
     ReferenceAcoshLayerTest::getTestCaseName);
+}  // namespace reference_tests
