+++ docs/template_plugin/tests/functional/op_reference/convert.cpp
@@ -12,6 +12,7 @@
 
 #include "base_reference_test.hpp"
 
+using namespace reference_tests;
 using namespace ngraph;
 using namespace InferenceEngine;
